+++ util.js
@@ -1,0 +1,4 @@
+// Получение случайного числа от 0 до num
+export function getRandom(num) {
+    return Math.ceil(Math.random() * num);
+}
